ci: disable eslint for building
Disables eslint for running and building the web app.
Also fixes eslint pre-commit hook to not show warning on ignored files,
and to actually fix the code when run.

The reasoning is that a linting tool should not stop the app from
building. The linting will still be checked in the tests.